requirements.txt
Flask
requests
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
pytest